fastapi-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
watchdog==3.0.0
PyYAML==6.0.1
rendercv>=2.0.0